Change price type to MONEY

+++ db/schema/04_items.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS items CASCADE;
 CREATE TABLE items (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
-  price INTEGER  NOT NULL,
+  price MONEY NOT NULL,
   description TEXT,
   photo_url VARCHAR(255) NOT NULL,
 );
